pyproject.toml
[tool.poetry]
name = "ocean-spaces"
version = "0.1.0"
description = ""
authors = ["dawoodumar76 <dawoodumar76@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
poetry-dotenv = "^0.4.0"
boto3 = "^1.35.26"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
